Reference

Contents

Reference

Contents

Index

Index

MPCCModels.AbstractMPCCModel

MPCCModels.CCType

MPCCModels.MPCCModel

MPCCModels.MPCCModel

MPCCModels.comp_left

MPCCModels.comp_left!

MPCCModels.comp_res_left

MPCCModels.comp_res_left!

MPCCModels.comp_res_prod!

MPCCModels.comp_res_right

MPCCModels.comp_res_right!

MPCCModels.comp_residual

MPCCModels.comp_residual_product

MPCCModels.comp_residual_sum

MPCCModels.comp_right

MPCCModels.comp_right!

MPCCModels.get_c_lin

MPCCModels.get_c_nln

MPCCModels.get_cc_l

MPCCModels.get_cc_r

MPCCModels.get_cc_types

MPCCModels.get_comp_left_nnzj

MPCCModels.get_comp_right_nnzj

MPCCModels.get_ind_c

MPCCModels.get_ind_cc1

MPCCModels.get_ind_cc2

MPCCModels.get_ind_j_comp_left_row_map

MPCCModels.get_ind_j_comp_left_triplets

MPCCModels.get_ind_j_comp_right_row_map

MPCCModels.get_ind_j_comp_right_triplets

MPCCModels.get_ind_j_lin_row_map

MPCCModels.get_ind_j_lin_triplets

MPCCModels.get_ind_j_nln_row_map

MPCCModels.get_ind_j_nln_triplets

MPCCModels.get_ind_j_triplets

MPCCModels.get_ind_lift

MPCCModels.get_ind_lift_var

MPCCModels.get_ind_lin_lift

MPCCModels.get_ind_lin_lift_var

MPCCModels.get_ind_nln_lift

MPCCModels.get_ind_nln_lift_var

MPCCModels.get_ind_x

MPCCModels.get_lin_nlift

MPCCModels.get_ncc

MPCCModels.get_nlift

MPCCModels.get_nln_nlift

MPCCModels.get_nlp_meta

MPCCModels.get_parent

MPCCModels.jac_comp_left

MPCCModels.jac_comp_left_coord

MPCCModels.jac_comp_left_coord!

MPCCModels.jac_comp_left_structure

MPCCModels.jac_comp_left_structure!

MPCCModels.jac_comp_right

MPCCModels.jac_comp_right_coord

MPCCModels.jac_comp_right_coord!

MPCCModels.jac_comp_right_structure

MPCCModels.jac_comp_right_structure!

MPCCModels.lcomp_left

MPCCModels.lcomp_left!

MPCCModels.lcomp_right

MPCCModels.lcomp_right!

MPCCModels.vertical_form

NLPModels.cons

NLPModels.cons!

NLPModels.grad

NLPModels.grad!

NLPModels.hess

NLPModels.hess_coord

NLPModels.hess_coord

NLPModels.hess_coord!

NLPModels.hess_coord!

NLPModels.hess_structure

NLPModels.hess_structure!

NLPModels.jac

NLPModels.jac_coord

NLPModels.jac_coord!

NLPModels.jac_structure

NLPModels.jac_structure!

NLPModels.obj

NLPModels.objgrad!

MPCCModels.AbstractMPCCModel — Type

Abstract type for problems in the form:

min f(w)
s.t. lbc ≤ c(w) ≤ ubc
     lbG ≤ G(w) ⟂ H(w) ≥ lbH

MPCCModels.comp_left! — Method

Evaluates G(x) in place.

MPCCModels.comp_left — Method

Evaluates G(x).

MPCCModels.comp_res_left! — Method

Evaluates $G(x) - \ell_G$, in place.

MPCCModels.comp_res_left — Method

Evaluates $G(x) - \ell_G$.

MPCCModels.comp_res_prod! — Method

Evaluates $(G(x)-\ell_G)\odot(H(x)-\ell_H)$ in place.

MPCCModels.comp_res_right! — Method

Evaluates $H(x) - \ell_H$, in place.

MPCCModels.comp_res_right — Method

Evaluates $H(x) - \ell_H$.

MPCCModels.comp_residual — Method

Evaluates $\vert\min(G(x)-\ell_G,H(x)-\ell_H)\vert_\infty$.

MPCCModels.comp_residual_product — Method

Evaluates $\vert (G(x)-\ell_G)\odot(H(x)-\ell_H)\vert_\infty$.

MPCCModels.comp_residual_sum — Method

Evaluates $(G(x)-\ell_G)\cdot(H(x)-\ell_H)$.

MPCCModels.comp_right! — Method

Evaluates H(x) in place.

MPCCModels.comp_right — Method

Evaluates H(x)

MPCCModels.jac_comp_left — Method

Evaluates $\nabla G(x)$.

MPCCModels.jac_comp_left_coord! — Method

Evaluates COO values of $\nabla G(x)$, in place.

MPCCModels.jac_comp_left_coord — Method

Evaluates COO values of $\nabla G(x)$.

MPCCModels.jac_comp_left_structure! — Method

Evaluates COO structure of $\nabla G(x)$, in place.

MPCCModels.jac_comp_left_structure — Method

Evaluates COO structure of $\nabla G(x)$.

MPCCModels.jac_comp_right — Method

Evaluates $\nabla H(x)$.

MPCCModels.jac_comp_right_coord! — Method

Evaluates COO values of $\nabla G(x)$, in place.

MPCCModels.jac_comp_right_coord — Method

Evaluates COO values of $\nabla H(x)$.

MPCCModels.jac_comp_right_structure! — Method

Evaluates COO structure of $\nabla H(x)$, in place.

MPCCModels.jac_comp_right_structure — Method

Evaluates COO structure of $\nabla H(x)$.

MPCCModels.lcomp_left! — Method

Evaluates $\ell_G$ in place.

MPCCModels.lcomp_left — Method

Evaluates $\ell_G$.

MPCCModels.lcomp_right! — Method

Evaluates $\ell_H$ in place.

MPCCModels.lcomp_right — Method

Evaluates $\ell_H$.

NLPModels.cons! — Method

Evaluates the constraints $c(x)$ excluding any which appear in $G(x)$ or $H(x)$, in place.

NLPModels.cons — Method

Evaluates the constraints $c(x)$ excluding any which appear in $G(x)$ or $H(x)$, to a newly allocated vector.

NLPModels.grad! — Method

Evaluates in place the gradient of the MPCC objective $\nabla f(x)$

NLPModels.grad — Method

Evaluates the gradient of the MPCC objective $\nabla f(x)$ to a newly allocated vector.

NLPModels.hess — Method

Evaluates the Hessian of the Lagrangian, without the complementarity contribution.

Assumes zero multipliers.

NLPModels.hess_coord! — Method

Evaluates the COO structure of the Hessian of the Lagrangian, without the complementarity contribution, in place.

NLPModels.hess_coord! — Method

Evaluates the COO structure of the Hessian of the Lagrangian, without the complementarity contribution, in place.

Assumes zero multipliers.

NLPModels.hess_coord — Method

Evaluates the COO values of the Hessian of the Lagrangian, without the complementarity contribution, to newly allocated vectors.

NLPModels.hess_coord — Method

Evaluates the COO values of the Hessian of the Lagrangian, without the complementarity contribution, to newly allocated vectors.

Assumes zero multipliers.

NLPModels.hess_structure! — Method

Evaluates the COO structure of the Hessian of the Lagrangian, without the complementarity contribution, in place.

NLPModels.hess_structure — Method

Evaluates the COO structure of the Hessian of the Lagrangian, without the complementarity contribution, to newly allocated vectors.

NLPModels.jac — Method

Evaluates the constraint jacobian $\nabla c(x)$ excluding any constraints which appear in $G(x)$ or $H(x)$.

NLPModels.jac_coord! — Method

Evaluates the COO values of jacobian $\nabla c(x)$ excluding any constraints which appear in $G(x)$ or $H(x)$, in place.

NLPModels.jac_coord — Method

Evaluates the COO values of jacobian $\nabla c(x)$ excluding any constraints which appear in $G(x)$ or $H(x)$, to a newly allocated vector.

NLPModels.jac_structure! — Method

Evaluates the COO structure of jacobian $\nabla c(x)$ excluding any constraints which appear in $G(x)$ or $H(x)$, in place.

NLPModels.jac_structure — Method

Evaluates the COO structure of jacobian $\nabla c(x)$ excluding any constraints which appear in $G(x)$ or $H(x)$, to newly allocated vectors.

NLPModels.obj — Method

Evaluates the MPCC objective $f(x)$

NLPModels.objgrad! — Method

Evaluates in place the gradient of the MPCC objective $\nabla f(x)$ and returns the objective $f(x)$.

MPCCModels.MPCCModel — Type

Concrete type for problems in the form:

min f(x)
s.t. lbc ≤ c(x) ≤ ubc
     lbG ≤ G(x) ⟂ H(x) ≥ lbH

where G(x) and H(x) are defined by index sets into the x and c(x) of an nlp of the from:

min f(x)
s.t. lbc ≤ c(x) ≤ ubc

MPCCModels.MPCCModel — Method

Constructor for MPCCModel in the form:

 min f(x)
 s.t. lbc ≤ c(w) ≤ ubc
      lbx₁ ≤ x₁ ⟂x₂ ≥ lbx₂

where x₁ and x₂ are defined by index sets indvcc1 and indvcc2.

MPCCModels.get_ind_lift — Method

        get_ind_lift(nlp)
        get_ind_lift(meta)
        Return the value ind_lift from meta or nlp.meta.

MPCCModels.get_ind_lift_var — Method

        get_ind_lift_var(nlp)
        get_ind_lift_var(meta)
        Return the value ind_lift_var from meta or nlp.meta.

MPCCModels.get_ind_lin_lift — Method

        get_ind_lin_lift(nlp)
        get_ind_lin_lift(meta)
        Return the value ind_lin_lift from meta or nlp.meta.

MPCCModels.get_ind_lin_lift_var — Method

        get_ind_lin_lift_var(nlp)
        get_ind_lin_lift_var(meta)
        Return the value ind_lin_lift_var from meta or nlp.meta.

MPCCModels.get_ind_nln_lift — Method

        get_ind_nln_lift(nlp)
        get_ind_nln_lift(meta)
        Return the value ind_nln_lift from meta or nlp.meta.

MPCCModels.get_ind_nln_lift_var — Method

        get_ind_nln_lift_var(nlp)
        get_ind_nln_lift_var(meta)
        Return the value ind_nln_lift_var from meta or nlp.meta.

MPCCModels.get_lin_nlift — Method

        get_lin_nlift(nlp)
        get_lin_nlift(meta)
        Return the value lin_nlift from meta or nlp.meta.

MPCCModels.get_nlift — Method

        get_nlift(nlp)
        get_nlift(meta)
        Return the value nlift from meta or nlp.meta.

MPCCModels.get_nln_nlift — Method

        get_nln_nlift(nlp)
        get_nln_nlift(meta)
        Return the value nln_nlift from meta or nlp.meta.

MPCCModels.get_parent — Method

        get_parent(nlp)
        get_parent(meta)
        Return the value parent from meta or nlp.meta.

MPCCModels.vertical_form — Method

"Verticalize" a generic MPCC by lifting the nonscalar compenents of $G(x)$ and $H(x)$.

MPCCModels.CCType — Type

Enum for types of complementarity constraints

MPCCModels.get_c_lin — Method

  get_c_lin(nlp)
  get_c_lin(meta)
  Return the value c_lin from meta or nlp.meta.

MPCCModels.get_c_nln — Method

  get_c_nln(nlp)
  get_c_nln(meta)
  Return the value c_nln from meta or nlp.meta.

MPCCModels.get_cc_l — Method

  get_cc_l(nlp)
  get_cc_l(meta)
  Return the value cc_l from meta or nlp.meta.

MPCCModels.get_cc_r — Method

  get_cc_r(nlp)
  get_cc_r(meta)
  Return the value cc_r from meta or nlp.meta.

MPCCModels.get_cc_types — Method

  get_cc_types(nlp)
  get_cc_types(meta)
  Return the value cc_types from meta or nlp.meta.

MPCCModels.get_comp_left_nnzj — Method

  get_comp_left_nnzj(nlp)
  get_comp_left_nnzj(meta)
  Return the value comp_left_nnzj from meta or nlp.meta.

MPCCModels.get_comp_right_nnzj — Method

  get_comp_right_nnzj(nlp)
  get_comp_right_nnzj(meta)
  Return the value comp_right_nnzj from meta or nlp.meta.

MPCCModels.get_ind_c — Method

  get_ind_c(nlp)
  get_ind_c(meta)
  Return the value ind_c from meta or nlp.meta.

MPCCModels.get_ind_cc1 — Method

  get_ind_cc1(nlp)
  get_ind_cc1(meta)
  Return the value ind_cc1 from meta or nlp.meta.

MPCCModels.get_ind_cc2 — Method

  get_ind_cc2(nlp)
  get_ind_cc2(meta)
  Return the value ind_cc2 from meta or nlp.meta.

MPCCModels.get_ind_j_comp_left_row_map — Method

  get_ind_j_comp_left_row_map(nlp)
  get_ind_j_comp_left_row_map(meta)
  Return the value ind_j_comp_left_row_map from meta or nlp.meta.

MPCCModels.get_ind_j_comp_left_triplets — Method

  get_ind_j_comp_left_triplets(nlp)
  get_ind_j_comp_left_triplets(meta)
  Return the value ind_j_comp_left_triplets from meta or nlp.meta.

MPCCModels.get_ind_j_comp_right_row_map — Method

  get_ind_j_comp_right_row_map(nlp)
  get_ind_j_comp_right_row_map(meta)
  Return the value ind_j_comp_right_row_map from meta or nlp.meta.

MPCCModels.get_ind_j_comp_right_triplets — Method

  get_ind_j_comp_right_triplets(nlp)
  get_ind_j_comp_right_triplets(meta)
  Return the value ind_j_comp_right_triplets from meta or nlp.meta.

MPCCModels.get_ind_j_lin_row_map — Method

  get_ind_j_lin_row_map(nlp)
  get_ind_j_lin_row_map(meta)
  Return the value ind_j_lin_row_map from meta or nlp.meta.

MPCCModels.get_ind_j_lin_triplets — Method

  get_ind_j_lin_triplets(nlp)
  get_ind_j_lin_triplets(meta)
  Return the value ind_j_lin_triplets from meta or nlp.meta.

MPCCModels.get_ind_j_nln_row_map — Method

  get_ind_j_nln_row_map(nlp)
  get_ind_j_nln_row_map(meta)
  Return the value ind_j_nln_row_map from meta or nlp.meta.

MPCCModels.get_ind_j_nln_triplets — Method

  get_ind_j_nln_triplets(nlp)
  get_ind_j_nln_triplets(meta)
  Return the value ind_j_nln_triplets from meta or nlp.meta.

MPCCModels.get_ind_j_triplets — Method

  get_ind_j_triplets(nlp)
  get_ind_j_triplets(meta)
  Return the value ind_j_triplets from meta or nlp.meta.

MPCCModels.get_ind_x — Method

  get_ind_x(nlp)
  get_ind_x(meta)
  Return the value ind_x from meta or nlp.meta.

MPCCModels.get_ncc — Method

  get_ncc(nlp)
  get_ncc(meta)
  Return the value ncc from meta or nlp.meta.

MPCCModels.get_nlp_meta — Method

  get_nlp_meta(nlp)
  get_nlp_meta(meta)
  Return the value nlp_meta from meta or nlp.meta.